requirements.txt
spotipy==2.19.0
pandas==1.3.0
numpy==1.21.2
pickle-mixin==1.0.2
Flask==2.0.1
sklearn==0.0
ibm-cloud-sdk-core==3.11.3
ibm-watson==5.2.3
python-dotenv==0.19.0
waitress==2.0.0
flask_json==0.3.4
